Split h.indexer into a package
To resolve a circular import in a future commit. `h.tasks.indexer` needs
to import the `SETTING_NEW_INDEX` constant, but `h.indexer` needed to
import `h.tasks.indexer` for the event subscriber.
